Split the Table as a reusable component
