+++ frontend/src/components/Admin/PagePermissionDashboard.js
@@ -1,54 +1,24 @@
 import React, { useState, useEffect } from "react";
 
-import { makeStyles } from "@material-ui/core/styles";
-import Paper from "@material-ui/core/Paper";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
-import TablePagination from "@material-ui/core/TablePagination";
-import TableRow from "@material-ui/core/TableRow";
-
 import { handleErrors, createProjectIDArray } from "utils/Utils";
 import { useAuth0 } from "components/common/ReactAuth0SPA";
+import PermissionDashboard from "components/common/Admin/PermissionDashboard";
+import LoadingSpinner from "components/common/LoadingSpinner";
 import * as CONSTANTS from "constants/Constants";
 
-const useStyles = makeStyles({
-  root: {
-    width: "100%",
-  },
-  container: {
-    maxHeight: "85%",
-  },
-});
+import "assets/style/PermissionDashboard.css";
 
 const PagePermissionDashboard = () => {
   const { getTokenSilently } = useAuth0();
-
-  const classes = useStyles();
 
   const [userData, setUserData] = useState({});
   const [userOption, setUserOption] = useState([]);
 
   const [allPermission, setAllPermission] = useState([]);
-  const [tableHeaders, setTableHeaders] = useState([]);
+  const [tableHeaderData, setTableHeaderData] = useState([]);
 
   const [allGrantedPermission, setAllGrantedPermission] = useState({});
-  const [tableBody, setTableBody] = useState([]);
-
-  // Hooks for Pagination
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(+event.target.value);
-    setPage(0);
-  };
+  const [tableBodyData, setTableBodyData] = useState([]);
 
   /*
     Fetchig All projets we have from Project API to become a table's header
@@ -111,7 +81,7 @@ const PagePermissionDashboard = () => {
         });
       }
 
-      setTableHeaders(tempArray);
+      setTableHeaderData(tempArray);
     }
   }, [allPermission]);
 
@@ -207,7 +177,7 @@ const PagePermissionDashboard = () => {
         tempObj = {};
       }
 
-      setTableBody(tempArray);
+      setTableBodyData(tempArray);
     }
   }, [allGrantedPermission, allPermission, userOption]);
 
@@ -265,63 +235,16 @@ const PagePermissionDashboard = () => {
   }, [userData]);
 
   return (
-    <Paper className={classes.root}>
-      <TableContainer className={classes.container}>
-        <Table stickyHeader aria-label="sticky table">
-          <TableHead>
-            <TableRow>
-              {tableHeaders.map((header) => (
-                <TableCell key={header.id} align={"center"}>
-                  {header.label}
-                </TableCell>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {tableBody
-              .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-              .map((eachUser) => {
-                return (
-                  <TableRow
-                    hover
-                    role="checkbox"
-                    tabIndex={-1}
-                    key={eachUser["auth0-user-id"]}
-                  >
-                    {tableHeaders.map((header) => {
-                      const value = eachUser[header.id];
-                      return (
-                        <TableCell
-                          key={header.id}
-                          align={"center"}
-                          style={
-                            header.id === "auth0-user-id"
-                              ? { backgroundColor: "white" }
-                              : value === "true"
-                              ? { backgroundColor: "green" }
-                              : { backgroundColor: "red" }
-                          }
-                        >
-                          {header.id === "auth0-user-id" ? value : null}
-                        </TableCell>
-                      );
-                    })}
-                  </TableRow>
-                );
-              })}
-          </TableBody>
-        </Table>
-      </TableContainer>
-      <TablePagination
-        rowsPerPageOptions={[10, 25]}
-        component="div"
-        count={tableBody.length}
-        rowsPerPage={rowsPerPage}
-        page={page}
-        onChangePage={handleChangePage}
-        onChangeRowsPerPage={handleChangeRowsPerPage}
-      />
-    </Paper>
+    <div className="permission-dashboard">
+      {tableHeaderData.length > 0 && tableBodyData.length > 0 ? (
+        <PermissionDashboard
+          tableHeaderData={tableHeaderData}
+          tableBodyData={tableBodyData}
+        />
+      ) : (
+        <LoadingSpinner className="permission-dashboard-loading-spinner" />
+      )}
+    </div>
   );
 };
